move general hotkeys to own component
