+++ src/components/Board/Board.tsx
@@ -6,6 +6,7 @@ import {ReactComponent as LeftArrowIcon} from "assets/icon-arrow-previous.svg";
 import {MenuBars} from "components/MenuBars";
 import {BoardHeader} from "components/BoardHeader";
 import "./Board.scss";
+import {HotkeyAnchor} from "components/HotkeyAnchor";
 
 export interface BoardProps {
   children: React.ReactElement<ColumnProps> | React.ReactElement<ColumnProps>[];
@@ -124,6 +125,7 @@ export const BoardComponent = ({children, currentUserIsModerator, moderating}: B
         <style>{`.board { --board__columns: ${columnsCount} }`}</style>
         <BoardHeader currentUserIsModerator={currentUserIsModerator} />
         <MenuBars />
+        <HotkeyAnchor />
         <main className="board" ref={boardRef}>
           {/* Fixed color - can also be dynamic */}
           <div className={`board__spacer-left ${getColorClassName("backlog-blue")}`} />
@@ -153,6 +155,7 @@ export const BoardComponent = ({children, currentUserIsModerator, moderating}: B
 
       <BoardHeader currentUserIsModerator={currentUserIsModerator} />
       <MenuBars />
+      <HotkeyAnchor />
 
       {state.showPreviousButton && (
         <button className={`board__navigation board__navigation-prev ${getColorClassName(columnColors[previousColumnIndex])}`} onClick={handlePreviousClick} aria-hidden>
