+++ src/components/HotkeyAnchor/HotkeyAnchor.scss
@@ -1,0 +1,3 @@
+.hotkey-anchor {
+  display: none;
+}
