+++ src/components/HotkeyAnchor/index.ts
@@ -1,0 +1,1 @@
+export * from "./HotkeyAnchor";
